refactor: display alert in dashboard page
