+++ src/pages/Dashboard/DashboardPage.vue
@@ -17,6 +17,9 @@
             <router-view @popUp="toggle_popup"></router-view>
         </div>
         <SummaryPopUp v-if="popup" @popUp="toggle_popup" :summaryID="summaryID" :card_name="card_name" class="absolute"/>
+        <!-- Alert -->
+        <Alert v-if="alert_status.status" :message="alert_status.message"
+               @closeAlert="close_alert"/>
     </section>
     <InformationPanel/>
 </template>
@@ -25,18 +28,30 @@
 
 import SummaryPopUp from "@/components/Summary/SummaryPopUp";
 import InformationPanel from "@/components/InformationPanel/InformationPanel";
+import Alert from "@/components/UI/Alert";
 
 export default {
     name: "DashboardPage",
     components: {
         InformationPanel,
-        SummaryPopUp
+        SummaryPopUp,
+        Alert
     },
     methods: {
         toggle_popup: function ({card_name, summaryID}) {
             this.summaryID = summaryID;
             this.card_name = card_name;
             this.popup = !this.popup;
+        },
+        close_alert: function(){
+            // Clears toggle alert timeout if alert is dismissed by the user
+            clearTimeout(this.$store.state.alert.timeout);
+            this.$store.state.alert.status = false;
+        }
+    },
+    computed: {
+        alert_status(){
+            return this.$store.state.alert;
         }
     },
     data() {
